+++ apps/web/src/data/graphql/requests/tokenQuery.ts
@@ -3,7 +3,7 @@ import isUndefined from 'lodash/isUndefined'
 import { CHAIN } from 'src/constants/network'
 import { sdk } from 'src/data/graphql/client'
 import { ImageMediaEncodingFragment } from 'src/data/graphql/sdk.generated'
-import { Token, TokenWinner } from 'src/typings'
+import { Token, TokenWinner } from "../../../modules/dao";
 
 export const tokenQuery = async (
   address: string,
