+++ apps/web/src/modules/create/components/ReviewAndDeploy/SuccessfulDeploy.tsx
@@ -13,10 +13,10 @@ import {
 } from 'src/styles/deploy.css'
 import { walletSnippet } from 'src/utils/helpers'
 import { transformFileProperties } from 'src/utils/transformFileProperties'
-import type { DaoContractAddresses } from 'src/typings'
 import { useContractRead } from 'wagmi'
 import { tokenAbi } from 'src/data/contract/abis'
 import { useLayoutStore } from 'src/stores'
+import { DaoContractAddresses } from "../../../dao";
 
 interface DeployedDaoProps extends DaoContractAddresses {
   title: string
