+++ apps/web/src/modules/dao/index.ts
@@ -1,3 +1,4 @@
 export * from './hooks'
 export * from './utils'
 export * from './components'
+export * from './typings'
