+++ apps/web/src/modules/transaction-builder/components/TransactionForm/forms/CustomTransaction/FormHandler.tsx
@@ -2,8 +2,8 @@ import React, { ReactElement } from 'react'
 import { AnimatePresence, motion } from 'framer-motion'
 import { Box } from '@zoralabs/zord'
 import { useFormStore } from 'src/stores/useFormStore'
-import { AddTransactionSection } from 'src/typings'
 import { useCustomTransactionStore } from '../../../../stores'
+import { AddTransactionSection } from "../../../../typings";
 
 interface FormHandler {
   forms: ReactElement[]
