+++ apps/web/src/modules/transaction-builder/index.ts
@@ -3,3 +3,4 @@ export * from './constants'
 export * from './hooks'
 export * from './stores'
 export * from './utils'
+export * from './typings'
