+++ apps/web/src/typings.ts
@@ -1,0 +1,27 @@
+import { MarketSortKey, Support } from 'src/data/graphql/sdk.generated'
+
+export { Support }
+
+export interface ExplorePageData {
+  daos: {
+    tokenId: string | null
+    collectionName: string | null
+    collectionAddress: string | null
+    name: string | null
+    image: string | null
+    endTime: number | null
+    highestBidder: string | null
+    highestBidPrice: number | null
+  }[]
+  pageInfo: {
+    limit: number
+    hasNextPage: boolean
+    endCursor?: string | null
+  }
+}
+
+export { MarketSortKey }
+
+export type AddressType = `0x${string}`
+
+export type BytesType = `0x${string}`
